Small changes
Issue #1 partly implemented

+++ second-app/module1/ejb/src/main/java/com/pohancenik/ws/SimpleWebService.java
@@ -19,6 +19,6 @@ public interface SimpleWebService {
 	 * @return
 	 */
 	@WebMethod
-	int getSimpleIntegerInRange(byte range);
+	int getSimpleIntegerInRange(int range);
 	
 }
